section 7: restructure client sources
