+++ Section7/src/common/ProfileHeader.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { graphql } from 'react-apollo/index';
+import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 import {
   Container,
