--- Section7/src/components/Constants.js
@@ -1,3 +1,0 @@
-export const ItemTypes = {
-  CARD: 'card',
-};
